Rename _utils.py to _materialize.py
- It now only contains the _materialize function, which can hardly be moved to another file since grad, jac and stack depend on it.
