+++ src/torchjd/autojac/_transform/jac.py
@@ -7,7 +7,7 @@ import torch
 from torch import Size, Tensor
 
 from ._differentiate import _Differentiate
-from ._utils import _materialize
+from ._materialize import _materialize
 from .tensor_dict import Jacobians
